+++ src/audio/manager.hpp
@@ -111,7 +111,6 @@ private:
     util::sync::AtomicBool recordActive = false;
     util::sync::AtomicBool recordQueuedStop = false;
     util::sync::AtomicBool recordQueuedHalt = false;
-    util::sync::AtomicBool recordStartDeferred = false;
     util::sync::AtomicBool recordingRaw = false;
     FMOD::Sound* recordSound = nullptr;
     size_t recordChunkSize = 0;