+++ src/audio/voice_record_manager.cpp
@@ -1,0 +1,76 @@
+#include "voice_record_manager.hpp"
+
+#include <data/packets/client/misc.hpp>
+#include <data/packets/client/game.hpp>
+#include <managers/error_queues.hpp>
+#include <net/network_manager.hpp>
+
+VoiceRecordingManager::VoiceRecordingManager() {
+    thread.setName("Audio Record Thread");
+    thread.setLoopFunction(&VoiceRecordingManager::threadFunc);
+    thread.start(this);
+}
+
+void VoiceRecordingManager::startRecording() {
+    queuedStart = true;
+}
+
+void VoiceRecordingManager::stopRecording() {
+    queuedStop = true;
+}
+
+void VoiceRecordingManager::threadFunc() {
+    auto& vm = GlobedAudioManager::get();
+
+    if (queuedStop) {
+        if (vm.isRecording()) {
+            vm.stopRecording();
+        }
+
+        this->resetBools(false);
+        return;
+    }
+
+    if (queuedStart) {
+        if (!vm.isRecording()) {
+            vm.validateDevices();
+
+            // make sure the recording device is valid
+            if (!vm.isRecordingDeviceSet()) {
+                ErrorQueues::get().warn("Unable to record audio, no recording device is set");
+                this->resetBools(false);
+                return;
+            }
+
+            auto result = vm.startRecording([](const auto& frame) {
+                auto& nm = NetworkManager::get();
+                if (!nm.established()) return;
+
+                // `frame` does not live long enough and will be destructed at the end of this callback.
+                // so we can't pass it directly in a `VoicePacket` and we use a `RawPacket` instead.
+
+                ByteBuffer buf;
+                buf.writeValue(frame);
+
+                nm.send(RawPacket::create(VoicePacket::PACKET_ID, VoicePacket::ENCRYPTED, VoicePacket::SHOULD_USE_TCP, std::move(buf)));
+            });
+
+            if (result.isErr()) {
+                ErrorQueues::get().warn(result.unwrapErr());
+                log::warn("unable to record audio: {}", result.unwrapErr());
+                this->resetBools(false);
+                return;
+            }
+        }
+
+        this->resetBools(true);
+    }
+
+    std::this_thread::sleep_for(util::time::millis(10));
+}
+
+void VoiceRecordingManager::resetBools(bool recording) {
+    this->recording = recording;
+    queuedStart = false;
+    queuedStop = false;
+}